+++ ktrain/utils.py
@@ -43,10 +43,18 @@ PREPROC_NAME = MODEL_BASENAME+'.preproc'
 #------------------------------------------------------------------------------
 
 def loss_fn_from_model(model):
+    # dep_fix
     if version.parse(tf.__version__) < version.parse('2.2'):
         return model.loss_functions[0].fn
     else: # TF 2.2.0
         return model.compiled_loss._get_loss_object(model.compiled_loss._losses[0].name).fn
+
+def metrics_from_model(model):
+    # dep_fix
+    if version.parse(tf.__version__) < version.parse('2.2') or DISABLE_V2_BEHAVIOR:
+        return [m.name for m in model.metrics] if is_tf_keras() else model.metrics
+    else: # TF 2.2.0
+        return model.compiled_metrics._metrics
 
 
 def is_classifier(model):
@@ -173,13 +181,27 @@ def is_imageclass_from_data(data):
     return type(data).__name__ in ['DirectoryIterator', 'DataFrameIterator', 'NumpyArrayIterator']
 
 
+def is_regression_from_data(data):
+    """
+    checks for regression task from data
+    """
+    data_arg_check(val_data=data, val_required=True)
+    if is_ner(data=data): return False          # NERSequence
+    elif is_nodeclass(data=data): return False  # NodeSequenceWrapper
+    elif is_linkpred(data=data): return False   #LinkSequenceWrapper
+    Y = y_from_data(data)
+    if len(Y.shape) == 1 or (len(Y.shape) > 1 and Y.shape[1] == 1): return True
+    return False
+
+
 def is_multilabel(data):
     """
     checks for multilabel from data
     """
     data_arg_check(val_data=data, val_required=True)
-    if is_ner(data=data): return False   # NERSequence
+    if is_ner(data=data): return False          # NERSequence
     elif is_nodeclass(data=data): return False  # NodeSequenceWrapper
+    elif is_linkpred(data=data): return False   #LinkSequenceWrapper
     multilabel = False
     Y = y_from_data(data)
     if len(Y.shape) == 1 or (len(Y.shape) > 1 and Y.shape[1] == 1): return False